refactoring Long to integer

+++ accounts-service/src/main/java/com/accounts/repository/CustomerRepository.java
@@ -2,11 +2,12 @@ package com.accounts.repository;
 
 import com.accounts.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface CustomerRepository extends JpaRepository<Customer,Long> {
+public interface CustomerRepository extends JpaRepository<Customer,Integer> {
     Optional<Customer> findByEmail(String Email);
     Optional<Customer> findByMobileNumber(String mobileNumber);
 }
